+++ controleestoqueapi/src/main/java/com/estoque/controle/model/produto/MovimentacaoDTO.java
@@ -1,4 +1,4 @@
 package com.estoque.controle.model.produto;
 
-public record MovimentacaoDTO(int quantidade, String descricao, int produtoId, TipoMovimentacao tipoMovimentacao) {
+public record MovimentacaoDTO(int quantidade, String descricao, Long produtoId, TipoMovimentacao tipoMovimentacao) {
 }
