+++ controleestoqueapi/src/main/java/com/estoque/controle/repository/FornecedorRepository.java
@@ -5,10 +5,12 @@ import com.estoque.controle.model.fornecedor.Fornecedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer> {
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Long> {
 
     List<Fornecedor> findByNomeFornecedor(String nomeCliente);
 
     String findByCpfOuCnpj(String cpfOuCnpj);
+
 }
